requirements.txt
langchain-openai
langgraph
wikipedia
python-dotenv
